Update the built-in PSACTs on Rancher's start-up.
Rancher has two built-in PSACTs: `rancher-privilege` and `rancher-restricted`, and Rancher may add new namespaces to the exemption namespace list in the `rancher-restricted` PSACT in a newer release. Previously, Rancher didn't update those two PSACTs if they already exist, which means that the exemption namespace list in rancher-restricted never gets updated once it is created.

With this PR, Rancher will update the built-in PSACTs such that it preserves the user's additions to the exemptions and merges everything from the built-in template into the existing one, which means that any value that is from the built-in template but removed by users will be added back. This is to make sure Rancher and its components work properly.